Revert "[UNO] Add batch upload promo to account settings page"
This reverts commit c892adffadcea24ed2f1b7ad8c685e9977aa6f02.

Reason for revert: Causing build failure on win32 - https://ci.chromium.org/ui/p/chromium/builders/ci/win32-official/10953/overview

Original change's description:
> [UNO] Add batch upload promo to account settings page
>
> This add a new entry point for batch upload to the account settings
> which is visible if the user has local data for data types where they
> have account storage enabled.
>
> Screenshot: https://screenshot.googleplex.com/8ED2z2nhzNMjEfZ
>
> Bug: 417938666
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7177483
> Commit-Queue: Amelie Schneider <amelies@google.com>
> Cr-Commit-Position: refs/heads/main@{#1552011}

Bug: 417938666
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7210944
Auto-Submit: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Owners-Override: Stephen McGruer <smcgruer@chromium.org>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1552126}

+++ chrome/browser/ui/webui/batch_upload_promo/batch_upload_promo_handler.cc
@@ -7,7 +7,6 @@
 #include "base/check_deref.h"
 #include "base/functional/bind.h"
 #include "base/functional/callback_helpers.h"
-#include "base/numerics/safe_conversions.h"
 #include "chrome/browser/profiles/batch_upload/batch_upload_service.h"
 #include "chrome/browser/profiles/batch_upload/batch_upload_service_factory.h"
 #include "chrome/browser/profiles/profile.h"
@@ -21,12 +20,12 @@ namespace {
 void RunCallbackOnLocalDataDescriptionReceived(
     BatchUploadPromoHandler::GetBatchUploadPromoLocalDataCountCallback callback,
     std::map<syncer::DataType, syncer::LocalDataDescription> local_data_map) {
-  int32_t local_data_count = std::accumulate(
+  int64_t local_data_count = std::accumulate(
       local_data_map.begin(), local_data_map.end(), 0,
-      [](int32_t current_count,
+      [](int64_t current_count,
          std::pair<syncer::DataType, syncer::LocalDataDescription> local_data) {
-        return current_count + base::checked_cast<int32_t>(
-                                   local_data.second.local_data_models.size());
+        return current_count +
+               static_cast<int64_t>(local_data.second.local_data_models.size());
       });
 
   std::move(callback).Run(local_data_count);
@@ -55,7 +54,7 @@ BatchUploadPromoHandler::~BatchUploadPromoHandler() {
 }
 
 void BatchUploadPromoHandler::OnLocalDataCountChanged(
-    int32_t local_data_count) {
+    int64_t local_data_count) {
   page_->OnLocalDataCountChanged(local_data_count);
 }
